% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiery-package.R
\docType{package}
\name{fiery-package}
\alias{fiery}
\alias{fiery-package}
\title{fiery: A Lightweight and Flexible Web Framework}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A very flexible framework for building server side logic in R. The 
    framework is unopinionated when it comes to how HTTP requests and WebSocket
    messages are handled and supports all levels of app complexity; from serving
    static content to full-blown dynamic web-apps. Fiery does not hold your hand
    as much as e.g. the shiny package does, but instead sets you free to create
    your web app the way you want.
}
\details{
\code{fiery} is a lightweight and flexible framework for web servers build on top of
the \link[httpuv:httpuv-package]{httpuv} package. The framework is largely
event-based, letting the developer attach handlers to life-cycle events as
well as defining and triggering their own events. This approach to
development is common in JavaScript, but might feel foreign to R developers.
Thankfully it is a rather simple concept that should be easy to gradually
begin to use to greater and greater effect.

\strong{Read more:}
\itemize{
\item Creation of the server object, along with all its methods and fields, is
described in the documentation of the \code{\link{Fire}} class.
\item An overview of the event model, along with descriptions of the predefined
life-cycle events and custom events can be found in the \link{events}
documentation.
\item A description of the \code{fiery} plugin interface and how to develop your
own plugins is laid out in the \link{plugins} documentation
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/fiery}
  \item Report bugs at \url{https://github.com/thomasp85/fiery/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

}
