% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{random_port}
\alias{random_port}
\title{Select a random safe port}
\usage{
random_port()
}
\value{
An integer in the range 49152-65535
}
\description{
This is a small utility function to get random safe ports to run your
application on. It chooses a port within the range that cannot be registered
to IANA and thus is safe to assume are not in use.
}
\examples{
random_port()

}
\keyword{internal}
