% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiery-package.R
\docType{package}
\name{fiery-package}
\alias{fiery}
\alias{fiery-package}
\title{A Fiery Framework for R-Based Web Development}
\description{
Fiery is a lightweight and flexible framework for web servers build on top
of the \link{httpuv} package. The framework is largely event-based, letting
the developer attach handlers to life-cycle events as well as defining and
triggering their own events. This approach to development is common in 
JavaScript, but might feel foreign to R developers. Thankfully it is a rather
simple concept that should be easy to gradually begin to use to greater and
greater effect.
}
\seealso{
\code{\link{Fire}} for the main approach to build your server logic
}

